feat: expose nodejs api
