+++ lib/markdownBuilder.js
@@ -27,10 +27,10 @@ const { keyword } = require('./keywords');
 function build({
   header,
   links = {},
-  includeproperties = [],
+  includeProperties = [],
   rewritelinks = (x) => x,
-  exampleformat = 'json',
-  skipproperties = [],
+  exampleFormat = 'json',
+  skipProperties = [],
 } = {}) {
   const formats = {
     'date-time': {
@@ -325,7 +325,7 @@ function build({
    * @param {*} props
    */
   function makeproptable(props = {}, patternProps = {}, additionalProps, required, slugger) {
-    if (skipproperties.includes('proptable')) {
+    if (skipProperties.includes('proptable')) {
       return paragraph();
     }
     const proprows = Object
@@ -365,7 +365,7 @@ function build({
   }
 
   function makearrayfact(items, additional) {
-    if (skipproperties.includes('arrayfact')) {
+    if (skipProperties.includes('arrayfact')) {
       return '';
     }
     return listItem([
@@ -468,17 +468,17 @@ function build({
       children.push(listItem(text(i18n`is optional`)));
     }
 
-    if (!skipproperties.includes('typefact')) {
+    if (!skipProperties.includes('typefact')) {
       children.push(maketypefact(definition));
     }
-    if (!skipproperties.includes('nullablefact')) {
+    if (!skipProperties.includes('nullablefact')) {
       children.push(makenullablefact(definition));
     }
-    if (!skipproperties.includes('definedinfact')) {
+    if (!skipProperties.includes('definedinfact')) {
       children.push(makedefinedinfact(definition));
     }
 
-    const additionalfacts = includeproperties.map((propname) => {
+    const additionalfacts = includeProperties.map((propname) => {
       if (definition[propname]) {
         return listItem(text(`${propname}: ${String(definition[propname])}`));
       }
@@ -534,7 +534,7 @@ function build({
   }
 
   function maketypesection(schema, level = 1) {
-    if (skipproperties.includes('typesection')) {
+    if (skipProperties.includes('typesection')) {
       return '';
     }
     const { children } = maketypefact(schema);
@@ -671,13 +671,13 @@ function build({
   }
 
   function makeexamples(schema, level = 1) {
-    if (schema[keyword`examples`] && schema[keyword`examples`].length > 0 && exampleformat === 'yaml') {
+    if (schema[keyword`examples`] && schema[keyword`examples`].length > 0 && exampleFormat === 'yaml') {
       return [
         heading(level + 1, text(i18n`${simpletitle(schema)} Examples`)),
         ...schema[keyword`examples`].map((example) => paragraph(code('yaml', yaml.dump(example, undefined, 2)))),
       ];
     }
-    if (schema[keyword`examples`] && schema[keyword`examples`].length > 0 && exampleformat === 'json') {
+    if (schema[keyword`examples`] && schema[keyword`examples`].length > 0 && exampleFormat === 'json') {
       return [
         heading(level + 1, text(i18n`${simpletitle(schema)} Examples`)),
         ...schema[keyword`examples`].map((example) => paragraph(code('json', JSON.stringify(example, undefined, 2)))),
